ansible-ml-2/ddp_train_example.requirements.txt
pytorch-lightning==0.7.6
hydra-core==0.11.3
torch==1.4.0
torchvision==0.5.0